requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
yarl==1.8.2
